StickyEvent unregister/register delivery bug
This issue exists in this case:
We have class A and its subclass B.
In
class BaseClass we have method "onEvent(A a)".
When we execute method
"postStick(new B())".
If BaseClass was registered, then its onEvent
method executes (it is
ok).
But if this class wasn't registered or call
unregister and then
registerSticky, then it will not receive event at
all. This is bug.