Specify SubscriberIndex size when initializing HashMap
Initializing the HashMap with a specific size is an optimization for when we're adding a large number of subscribers to the index. By default the HashMap will only be created with a capacity of 4, so if we add a large number of subscribers the map will have to expand it's size several times. This can cause garbage collections that slow down App launch.