Fix for crash on AndroidX
Eventbus is crashing when my app is running on a device with an API level below 26 and using the AndroidX libraries.

The problem is that a NoClassDefFoundError is being thrown for the TextClassifier that's a parameter in the AppCompatTextView#setTextClassifier() method. The reflection code you use doesn't filter for API level, unfortunately.

A quick fix is to add "androidx." to the system classes in the moveToSuperClass() method. You should probably do this for performance reasons in any case.