Refinements to the class not found exception fix
The EventBus API will crash if it runs into a method that takes a parameter that the app doesn't recognise. Android gets around this using the @TargetAPI annotation, but this requires a minimum API level of 15.

The easiest way to fix this is reverse the order of checking the number of parameters a method takes and whether the Subscribe annotation is present. This means the getParameterTypes() call is deferred, and if a system library incorporates something not available on a given API level, it won't find an unknown type.