[BREAD-2162] Do not fail when eventBusIndex is not present
EventBus annotation processor is conflicting with hilt annotation
processing which results in a broken incremental compilation.

Hilt runs annotation processing in a special gradle task that also
happens to execute other annotation processors when kapt task is
up-to-date. But hilt does not pass the original arguments to the
processors which eventbus expects to see.

This change changes behavior to not fail annotation processing but to
instead not run it.
This should be safe as:
 - we do have `eventBusIndex` configured in our projects and don't
 really need to eventbus to double-check this
 - the problem only occurs when EventBusAnnotationProcessor would not
 have been executed as kapt task is up-to-date and is only executed
 because hilt forces it to execute.